Added Cypress Test For OrganizationDetails
