+++ src/people/widgetViews/__tests__/OrganizationDetails.spec.tsx
@@ -1,17 +1,23 @@
 import '@testing-library/jest-dom';
-import '@testing-library/jest-dom/extend-expect';
-import { BrowserRouter as Router } from 'react-router-dom';
-import { act, render, fireEvent, waitFor } from '@testing-library/react';
-import { person } from '__test__/__mockData__/persons';
+import { act, render, waitFor, within, screen, fireEvent } from '@testing-library/react';
+import { people, person } from '__test__/__mockData__/persons';
 import { user } from '__test__/__mockData__/user';
 import nock from 'nock';
 import React from 'react';
-import { Organization, mainStore } from 'store/main';
+import { Organization, PaymentHistory, mainStore } from 'store/main';
+import { uiStore } from 'store/ui';
+import * as LighningDecoder from 'light-bolt11-decoder';
+import { BrowserRouter as Router, MemoryRouter } from 'react-router-dom';
 import OrganizationDetails from '../OrganizationDetails';
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
   useRouteMatch: () => ({ url: '', path: '' })
+}));
+
+jest.mock('light-bolt11-decoder', () => ({
+  ...jest.requireActual('light-bolt11-decoder'),
+  decode: jest.fn()
 }));
 
 const organization: Organization = {
@@ -24,21 +30,697 @@ const organization: Organization = {
   uuid: 'c360e930-f94d-4c07-9980-69fc428a994e',
   bounty_count: 1,
   budget: 100000,
-  owner_pubkey: 'clrqpq84nncuuf32kh2g',
+  owner_pubkey: user.pubkey!,
   img: 'https://memes.sphinx.chat/public/3bt5n-7mGLgC6jGBBwKwLyZdJY6IUVZke8p2nLUsPhU=',
   created: '2023-12-12T00:44:25.83042Z',
   updated: '2023-12-12T01:12:39.970648Z',
   deleted: false
 };
 
-/**
- * @jest-environment jsdom
- */
+const orgUser: Organization = {
+  id: 'clrqpq84nncuuf32kh2g',
+  name: 'test organization',
+  description: 'test',
+  github: 'https://github.com/stakwork',
+  website: 'https://community.sphinx.chat',
+  show: true,
+  uuid: 'clu80datu2rjujsmim40',
+  bounty_count: 1,
+  budget: 10000,
+  owner_pubkey: user.pubkey!,
+  img: 'https://memes.sphinx.chat/public/3bt5n-7mGLgC6jGBBwKwLyZdJY6IUVZke8p2nLUsPhU=',
+  created: '2023-12-12T00:44:25.83042Z',
+  updated: '2023-12-12T01:12:39.970648Z',
+  deleted: false
+};
+
+const roles = [
+  {
+    name: 'VIEW REPORT'
+  }
+];
+
+const invoiceDetails = [
+  {
+    id: 1,
+    amount: 10,
+    bounty_id: 0,
+    payment_type: 'withdraw',
+    org_uuid: 'cn6pq4qtu2rj3nhh5kkg',
+    sender_pubkey: '029a49ce2ec0885bd5edb09dbc4e4f700529dd76a7d19baab1656da61d909304c1',
+    receiver_pubkey: '',
+    created: '2024-02-15T05:24:13.021248Z',
+    updated: '2024-02-15T05:24:13.021248Z',
+    status: true,
+    sender_name: 'test sender 1',
+    receiver_name: '',
+    sender_img: 'https://memes.sphinx.chat/public/o6xp_Ai9IdFEKMPNYE_QDF2MPoWHWuXn0pdlNgi12Gk=',
+    receiver_img: ''
+  },
+  {
+    id: 2,
+    amount: 100,
+    bounty_id: 0,
+    payment_type: 'deposit',
+    org_uuid: 'cn6pq4qtu2rj3nhh5kkg',
+    sender_pubkey: '029a49ce2ec0885bd5edb09dbc4e4f700529dd76a7d19baab1656da61d909304c1',
+    receiver_pubkey: '',
+    created: '2024-02-15T05:22:33.451Z',
+    updated: '2024-02-15T05:22:33.451Z',
+    status: true,
+    sender_name: 'test sender 2',
+    receiver_name: '',
+    sender_img: 'https://memes.sphinx.chat/public/o6xp_Ai9IdFEKMPNYE_QDF2MPoWHWuXn0pdlNgi12Gk=',
+    receiver_img: ''
+  },
+  {
+    id: 3,
+    amount: 10,
+    bounty_id: 0,
+    payment_type: 'withdraw',
+    org_uuid: 'cn6pq4qtu2rj3nhh5kkg',
+    sender_pubkey: '029a49ce2ec0885bd5edb09dbc4e4f700529dd76a7d19baab1656da61d909304c1',
+    receiver_pubkey: '',
+    created: '2024-02-15T05:24:13.021248Z',
+    updated: '2024-02-15T05:24:13.021248Z',
+    status: true,
+    sender_name: 'test sender 3',
+    receiver_name: 'receiver 3',
+    sender_img: 'https://memes.sphinx.chat/public/o6xp_Ai9IdFEKMPNYE_QDF2MPoWHWuXn0pdlNgi12Gk=',
+    receiver_img: ''
+  },
+  {
+    id: 4,
+    amount: 10,
+    bounty_id: 0,
+    payment_type: 'payment',
+    org_uuid: 'cn6pq4qtu2rj3nhh5kkg',
+    sender_pubkey: '029a49ce2ec0885bd5edb09dbc4e4f700529dd76a7d19baab1656da61d909304c1',
+    receiver_pubkey: '',
+    created: '2024-02-15T05:24:13.021248Z',
+    updated: '2024-02-15T05:24:13.021248Z',
+    status: true,
+    sender_name: 'test sender 4',
+    receiver_name: 'receiver 4',
+    sender_img: 'https://memes.sphinx.chat/public/o6xp_Ai9IdFEKMPNYE_QDF2MPoWHWuXn0pdlNgi12Gk=',
+    receiver_img: ''
+  }
+] as PaymentHistory[];
+
+let closeFn;
+let resetOrgFn;
+let getOrgFn;
+
+beforeEach(() => {
+  jest.spyOn(mainStore, 'getPersonAssignedBounties').mockReturnValue(Promise.resolve([]));
+  jest.spyOn(mainStore, 'pollOrgBudgetInvoices').mockReturnValue(Promise.resolve([]));
+  jest.spyOn(mainStore, 'organizationInvoiceCount').mockReturnValue(Promise.resolve(0));
+  jest.spyOn(mainStore, 'getOrganizationUsers').mockReturnValue(Promise.resolve(people));
+  jest.spyOn(mainStore, 'getUserRoles').mockReturnValue(Promise.resolve(roles));
+  mainStore.bountyRoles = roles;
+  uiStore.setMeInfo(user);
+  jest
+    .spyOn(mainStore, 'getOrganizationBudget')
+    .mockReturnValue(Promise.resolve({ total_budget: 10000 }));
+  jest.spyOn(mainStore, 'getPaymentHistories').mockReturnValue(Promise.resolve(invoiceDetails));
+  resetOrgFn = jest.fn();
+  closeFn = jest.fn();
+  getOrgFn = jest.fn();
+});
+
+afterEach(() => {
+  jest.resetAllMocks();
+});
 describe('OrganizationDetails', () => {
   jest.setTimeout(20000);
-  //nock.disableNetConnect();
+  nock.disableNetConnect();
   nock(user.url).get('/person/id/1').reply(200, { user });
   nock(user.url).get('/ask').reply(200, {});
+
+  it('Should render history button on organization page', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      expect(getByTestId('organization-view-transaction-history-button')).toBeInTheDocument();
+    });
+  });
+
+  it('Should open history modal on click button', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+    });
+  });
+
+  it('Should render correct number of transaction in history modal', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      for (let i = 0; i < invoiceDetails.length; i++) {
+        await waitFor(() => getByTestId(`payment-history-transaction-${i}`));
+      }
+    });
+  });
+
+  it('Should render correct status of a transaction', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(async () => getByTestId('payment-history-modal'));
+      await waitFor(async () => getByTestId(`payment-history-transaction-0`));
+      expect(
+        within(getByTestId('payment-history-transaction-0')).getByTestId(
+          'payment-history-transaction-type'
+        ).innerHTML
+      ).toBe('withdraw');
+
+      expect(
+        within(getByTestId('payment-history-transaction-1')).getByTestId(
+          'payment-history-transaction-type'
+        ).innerHTML
+      ).toBe('deposit');
+      expect(
+        within(getByTestId('payment-history-transaction-3')).getByTestId(
+          'payment-history-transaction-type'
+        ).innerHTML
+      ).toBe('payment');
+    });
+  });
+
+  it('Should render correct transaction amount of a transaction', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      await waitFor(() => getByTestId(`payment-history-transaction-0`));
+
+      expect(
+        within(getByTestId('payment-history-transaction-0')).getByTestId(
+          'payment-history-transaction-amount'
+        )
+      ).toHaveTextContent(`${invoiceDetails[0].amount} sats`);
+      expect(
+        within(getByTestId('payment-history-transaction-1')).getByTestId(
+          'payment-history-transaction-amount'
+        )
+      ).toHaveTextContent(`${invoiceDetails[1].amount} sats`);
+    });
+  });
+
+  it('Should render correct sender of a transaction', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      await waitFor(() => getByTestId(`payment-history-transaction-0`));
+
+      expect(
+        within(getByTestId('payment-history-transaction-0')).getByTestId(
+          'payment-history-transaction-sender'
+        ).innerHTML
+      ).toContain(invoiceDetails[0].sender_name);
+    });
+  });
+
+  it('Should render correct receiver of a transaction', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      expect(
+        within(getByTestId('payment-history-transaction-3')).getByTestId(
+          'payment-history-transaction-receiver'
+        ).innerHTML
+      ).toContain(invoiceDetails[3].receiver_name);
+    });
+  });
+
+  it('Should redirect to correct url on selecting a transaction', async () => {
+    const opeWindowFn = jest.fn();
+    global.open = opeWindowFn;
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      within(getByTestId('payment-history-transaction-3'))
+        .getByTestId('payment-history-transaction-link')
+        .click();
+      expect(opeWindowFn).toBeCalledWith(`/bounty/${invoiceDetails[3].bounty_id}`, '_blank');
+    });
+  });
+
+  it('Should render all types of filter in history modal ', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      getByTestId('payment-history-filter-type-payment');
+      getByTestId('payment-history-filter-type-deposit');
+      getByTestId('payment-history-filter-type-withdraw');
+    });
+  });
+
+  it('Should render transaction according to the selected filter', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      getByTestId('organization-view-transaction-history-button').click();
+      await waitFor(() => getByTestId('payment-history-modal'));
+      within(getByTestId('payment-history-filter-type-payment')).getByLabelText('Payments').click();
+      within(getByTestId('payment-history-filter-type-deposit')).getByLabelText('Deposit').click();
+
+      getByTestId('payment-history-filter-type-deposit');
+      getByTestId('payment-history-filter-type-withdraw');
+    });
+  });
+
+  it('Should renders settings and delete buttons', async () => {
+    render(
+      <MemoryRouter>
+        <OrganizationDetails
+          close={closeFn}
+          getOrganizations={getOrgFn}
+          org={orgUser}
+          resetOrg={resetOrgFn}
+        />
+      </MemoryRouter>
+    );
+    waitFor(async () => {
+      const settingButton = await screen.findByTestId('settings-icon');
+      expect(settingButton).toBeInTheDocument();
+
+      const deleteButton = await screen.findByTestId('delete-icon');
+      expect(deleteButton).toBeInTheDocument();
+    });
+  });
+
+  it('Should renders manage org, manage bounties, fund org, withdraw from org, view txn history, update members, user profile of member, close modal "X" button, and update roles button', async () => {
+    act(async () => {
+      render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={orgUser}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+      await waitFor(async () => {
+        // Test for Edit View Button
+        const editButton = await screen.findByTestId('edit-button');
+        expect(editButton).toBeInTheDocument();
+
+        // Test for Manage Bounties View Button
+        const viewBounty = await screen.findByTestId('view-bounties');
+        expect(viewBounty).toBeInTheDocument();
+
+        // Test for Transaction History View Button
+        const HistoryButton = await screen.findByTestId('history-button');
+        expect(HistoryButton).toBeInTheDocument();
+
+        // Test for Withdrawal Button
+        const withdrawButton = await screen.findByTestId('withdrawal-button');
+        expect(withdrawButton).toBeInTheDocument();
+
+        // Test for Deposit Button
+        const depositButton = await screen.findByTestId('deposit-button');
+        expect(depositButton).toBeInTheDocument();
+
+        // Verify Transaction History Text is Present
+        const txnHistory = await screen.getByText(/YOUR BALANCE/i);
+        expect(txnHistory).toBeInTheDocument();
+
+        // Verify Budget Display
+        const Budget = await screen.getByText(/10,000/i);
+        expect(Budget).toBeInTheDocument();
+
+        // Test for Adding a User Button
+        const addUserButton = await screen.findByTestId('add-user');
+        expect(addUserButton).toBeInTheDocument();
+
+        // Verify User Profile Display
+        const imageIcon = await screen.findByTestId('avatarIcon');
+        expect(imageIcon).toBeInTheDocument();
+
+        const name = await screen.findByTestId('user_alias');
+        expect(name).toBeInTheDocument();
+
+        const userPubkey = await screen.findByTestId('user_pubkey');
+        expect(userPubkey).toBeInTheDocument();
+
+        // Test for Opening Settings
+        const settingButton = await screen.findByTestId('settings-icon');
+        fireEvent.click(settingButton);
+
+        // Test for Updating Roles Button
+        const updateRoleButton = await screen.findByRole('button', { name: /Updates roles/i });
+        expect(updateRoleButton).toBeInTheDocument();
+
+        // Test for Closing Modal
+        const closeButton = await screen.findByTestId('close-btn');
+        expect(closeButton).toBeInTheDocument();
+      });
+    });
+  });
+
+  it('Allows selecting and deselecting user roles', async () => {
+    act(async () => {
+      render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={orgUser}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      await waitFor(async () => {
+        const settingButton = await screen.findByTestId('settings-icon');
+        fireEvent.click(settingButton);
+
+        // Verify the roles section is present
+        const rolePresent = await screen.getByText(/User Roles/i);
+        expect(rolePresent).toBeInTheDocument();
+
+        // Find and interact with a specific role checkbox
+        const roleCheckbox = await screen.findByLabelText('Manage bounties');
+        expect(roleCheckbox).not.toBeChecked();
+
+        // Select the role
+        fireEvent.click(roleCheckbox);
+        expect(roleCheckbox).toBeChecked();
+
+        // Deselect the role
+        fireEvent.click(roleCheckbox);
+        expect(roleCheckbox).not.toBeChecked();
+      });
+    });
+  });
+
+  it('Saves roles and shows a success notification upon clicking "update roles"', async () => {
+    act(async () => {
+      const addToast = jest.fn();
+      render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={orgUser}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      await waitFor(async () => {
+        const updateRoleButton = await screen.findByRole('button', { name: /Updates roles/i });
+        fireEvent.click(updateRoleButton);
+
+        expect(addToast).toHaveBeenCalledWith('Roles Updated Successfully', 'success');
+      });
+    });
+  });
+
+  it('Delete user and shows a success notification upon clicking "Delete Button"', async () => {
+    act(async () => {
+      const addToast = jest.fn();
+      render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={orgUser}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      await waitFor(async () => {
+        const updateRoleButton = await screen.findByRole('button', { name: /Delete/i });
+        fireEvent.click(updateRoleButton);
+
+        expect(addToast).toHaveBeenCalledWith('User deleted successfully', 'success');
+      });
+    });
+  });
+
+  it('Test that withdraw modal is rendered', async () => {
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      const withdrawBtn = getByTestId('organization-withdraw-budget-button');
+      expect(withdrawBtn).toBeInTheDocument();
+
+      fireEvent.click(withdrawBtn);
+
+      await waitFor(() => {
+        const modal = getByTestId('testid-modal');
+        expect(modal).toBeInTheDocument();
+
+        expect(within(modal).getByText('Withdraw')).toBeInTheDocument();
+        expect(within(modal).getByText('Paste your invoice')).toBeInTheDocument();
+        expect(within(modal).getByTestId('withdrawInvoiceInput')).toBeInTheDocument();
+      });
+    });
+  });
+
+  it('Test that user is able to post an invoice and can pays the invoice correctly', async () => {
+    const mockedDecode = LighningDecoder.decode as jest.MockedFunction<
+      typeof LighningDecoder.decode
+    >;
+    mockedDecode.mockReturnValue({
+      sections: [{ value: 1 }, { value: 2 }, { value: 1000000 }]
+    });
+
+    jest.spyOn(mainStore, 'withdrawBountyBudget').mockResolvedValue({
+      success: true,
+      response: {
+        success: true,
+        response: {
+          payment_request: ''
+        }
+      }
+    });
+
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      const withdrawBtn = getByTestId('organization-withdraw-budget-button');
+      expect(withdrawBtn).toBeInTheDocument();
+
+      fireEvent.click(withdrawBtn);
+
+      const modal = await waitFor(() => getByTestId('testid-modal'));
+
+      await waitFor(() => {
+        const invoice =
+          'lnbc20u1p3y0x3hpp5743k2g0fsqqxj7n8qzuhns5gmkk4djeejk3wkp64ppevgekvc0jsdqcve5kzar2v9nr5gpqd4hkuetesp5ez2g297jduwc20t6lmqlsg3man0vf2jfd8ar9fh8fhn2g8yttfkqxqy9gcqcqzys9qrsgqrzjqtx3k77yrrav9hye7zar2rtqlfkytl094dsp0ms5majzth6gt7ca6uhdkxl983uywgqqqqlgqqqvx5qqjqrzjqd98kxkpyw0l9tyy8r8q57k7zpy9zjmh6sez752wj6gcumqnj3yxzhdsmg6qq56utgqqqqqqqqqqqeqqjq7jd56882gtxhrjm03c93aacyfy306m4fq0tskf83c0nmet8zc2lxyyg3saz8x6vwcp26xnrlagf9semau3qm2glysp7sv95693fphvsp54l567';
+
+        expect(modal).toBeInTheDocument();
+
+        //  Test that user is able to post an invoice
+        fireEvent.change(within(modal).getByTestId('withdrawInvoiceInput'), {
+          target: { value: invoice }
+        });
+      });
+
+      await waitFor(() => {
+        const confirmBtn = within(modal).getByText('Confirm');
+
+        // Test invoice is correct
+        expect(confirmBtn).not.toBeDisabled();
+        fireEvent.click(confirmBtn);
+      });
+
+      await waitFor(() => {
+        expect(within(modal).getByText('You are about to withdraw')).toBeInTheDocument();
+        expect(within(modal).getByTestId('WithdrawAmount')).toHaveTextContent('1,000 SATS');
+      });
+
+      //  Test that confirm pays the invoice
+      const confirmPayBtn = within(modal).getByText('Withdraw');
+      expect(confirmPayBtn).not.toBeDisabled();
+      fireEvent.click(confirmPayBtn);
+
+      await waitFor(() => {
+        expect(within(modal).getByText('Successfully Withdraw')).toBeInTheDocument();
+      });
+    });
+  });
+
+  it('Test that error message appears if insufficient funds to withdraw', async () => {
+    const mockedDecode = LighningDecoder.decode as jest.MockedFunction<
+      typeof LighningDecoder.decode
+    >;
+    mockedDecode.mockReturnValue({
+      sections: [{ value: 1 }, { value: 2 }, { value: 1000000 }]
+    });
+
+    jest.spyOn(mainStore, 'withdrawBountyBudget').mockResolvedValue({
+      success: false,
+      error: 'Organization budget is not enough to withdraw the amount'
+    });
+
+    await act(async () => {
+      const { getByTestId } = render(
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
+      );
+
+      const withdrawBtn = getByTestId('organization-withdraw-budget-button');
+      expect(withdrawBtn).toBeInTheDocument();
+
+      fireEvent.click(withdrawBtn);
+
+      const modal = await waitFor(() => getByTestId('testid-modal'));
+
+      const invoice =
+        'lnbc20u1p3y0x3hpp5743k2g0fsqqxj7n8qzuhns5gmkk4djeejk3wkp64ppevgekvc0jsdqcve5kzar2v9nr5gpqd4hkuetesp5ez2g297jduwc20t6lmqlsg3man0vf2jfd8ar9fh8fhn2g8yttfkqxqy9gcqcqzys9qrsgqrzjqtx3k77yrrav9hye7zar2rtqlfkytl094dsp0ms5majzth6gt7ca6uhdkxl983uywgqqqqlgqqqvx5qqjqrzjqd98kxkpyw0l9tyy8r8q57k7zpy9zjmh6sez752wj6gcumqnj3yxzhdsmg6qq56utgqqqqqqqqqqqeqqjq7jd56882gtxhrjm03c93aacyfy306m4fq0tskf83c0nmet8zc2lxyyg3saz8x6vwcp26xnrlagf9semau3qm2glysp7sv95693fphvsp54l567';
+
+      expect(modal).toBeInTheDocument();
+
+      //  Test that user is able to post an invoice
+      fireEvent.change(within(modal).getByTestId('withdrawInvoiceInput'), {
+        target: { value: invoice }
+      });
+
+      await waitFor(() => {
+        const confirmBtn = within(modal).getByText('Confirm');
+        fireEvent.click(confirmBtn);
+      });
+
+      await waitFor(() => {
+        expect(within(modal).getByText('You are about to withdraw')).toBeInTheDocument();
+        expect(within(modal).getByTestId('WithdrawAmount')).toHaveTextContent('1,000 SATS');
+      });
+
+      const confirmPayBtn = within(modal).getByText('Withdraw');
+      expect(confirmPayBtn).not.toBeDisabled();
+      fireEvent.click(confirmPayBtn);
+
+      await waitFor(() => {
+        expect(
+          within(modal).getByText('Organization budget is not enough to withdraw the amount')
+        ).toBeInTheDocument();
+      });
+    });
+  });
 
   it('Add User button shows Add User modal', async () => {
     const props: {
@@ -85,6 +767,7 @@ describe('OrganizationDetails', () => {
       });
     })();
   });
+
   it('render organization name', async () => {
     jest.spyOn(mainStore, 'getPersonAssignedBounties').mockReturnValue(Promise.resolve([]));
     jest.spyOn(mainStore, 'pollOrgBudgetInvoices').mockReturnValue(Promise.resolve([]));
@@ -100,12 +783,14 @@ describe('OrganizationDetails', () => {
     const getOrgFn = jest.fn();
     await act(async () => {
       const { getByText } = render(
-        <OrganizationDetails
-          close={closeFn}
-          getOrganizations={getOrgFn}
-          org={organization}
-          resetOrg={resetOrgFn}
-        />
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
       );
 
       expect(getByText(organization.name)).toBeInTheDocument();
@@ -142,12 +827,14 @@ describe('OrganizationDetails', () => {
     const getOrgFn = jest.fn();
     await act(async () => {
       const { getByRole } = render(
-        <OrganizationDetails
-          close={closeFn}
-          getOrganizations={getOrgFn}
-          org={organization}
-          resetOrg={resetOrgFn}
-        />
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
       );
 
       const depositBtn = getByRole('button', { name: 'Deposit' });
@@ -176,21 +863,25 @@ describe('OrganizationDetails', () => {
     const closeFn = jest.fn();
     const resetOrgFn = jest.fn();
     const getOrgFn = jest.fn();
-    await act(async () => {
+    act(async () => {
       const { getByRole } = render(
-        <OrganizationDetails
-          close={closeFn}
-          getOrganizations={getOrgFn}
-          org={organization}
-          resetOrg={resetOrgFn}
-        />
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
       );
 
-      const editBtn = getByRole('button', { name: 'Edit' });
-      const addUsersBtn = getByRole('button', { name: 'Add User' });
+      await waitFor(() => {
+        const editBtn = getByRole('button', { name: 'Edit' });
+        const addUsersBtn = getByRole('button', { name: 'Add User' });
 
-      expect(editBtn).toBeDisabled();
-      expect(addUsersBtn).toBeDisabled();
+        expect(editBtn).toBeDisabled();
+        expect(addUsersBtn).toBeDisabled();
+      });
     });
   });
 
@@ -207,14 +898,16 @@ describe('OrganizationDetails', () => {
     const closeFn = jest.fn();
     const resetOrgFn = jest.fn();
     const getOrgFn = jest.fn();
-    await act(async () => {
+    act(async () => {
       const { getByRole } = render(
-        <OrganizationDetails
-          close={closeFn}
-          getOrganizations={getOrgFn}
-          org={{ ...organization, bounty_count: 0 }}
-          resetOrg={resetOrgFn}
-        />
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={{ ...organization, bounty_count: 0 }}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
       );
 
       const viewBountiesBtn = getByRole('button', { name: 'View Bounties open_in_new' });
@@ -244,21 +937,24 @@ describe('OrganizationDetails', () => {
     const closeFn = jest.fn();
     const resetOrgFn = jest.fn();
     const getOrgFn = jest.fn();
-    await act(async () => {
+    act(async () => {
       const { getByRole } = render(
-        <OrganizationDetails
-          close={closeFn}
-          getOrganizations={getOrgFn}
-          org={organization}
-          resetOrg={resetOrgFn}
-        />
+        <MemoryRouter>
+          <OrganizationDetails
+            close={closeFn}
+            getOrganizations={getOrgFn}
+            org={organization}
+            resetOrg={resetOrgFn}
+          />
+        </MemoryRouter>
       );
+      await waitFor(() => {
+        const withdrawBtn = getByRole('button', { name: 'Withdraw' });
+        const historyBtn = getByRole('button', { name: 'History' });
 
-      const withdrawBtn = getByRole('button', { name: 'Withdraw' });
-      const historyBtn = getByRole('button', { name: 'History' });
-
-      expect(withdrawBtn).toBeDisabled();
-      expect(historyBtn).toBeDisabled();
+        expect(withdrawBtn).toBeDisabled();
+        expect(historyBtn).toBeDisabled();
+      });
     });
   });
 });
